pyproject.toml
[tool.poetry]
name = "cyber-threat-ai"
version = "0.1.0"
description = "Production-ready AI-powered cyber threat detection system with multi-modal analysis and neuro-symbolic reasoning"
authors = ["Kunal SIngh"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.0"
torchvision = "^0.16.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.4.0"
loguru = "^0.7.2"
pyyaml = "^6.0.1"
networkx = "^3.2"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
transformers = "^4.35.0"
torch-geometric = "^2.4.0"
gymnasium = "^0.29.0"
stable-baselines3 = "^2.2.0"
captum = "^0.6.0"
pyarrow = "^14.0.0"
tqdm = "^4.66.0"
tensorboard = "^2.15.0"
omegaconf = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.9.0"
isort = "^5.12.0"
mypy = "^1.6.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
